bot/requirements.txt
python-telegram-bot
supabase
datetime
langgraph
langchain
openai
langchain_openai
jsonschema
litellm
groq
pypdf
python-docx
sentence-transformers
torch